+++ packages/porter-cli/History.md
@@ -1,4 +1,9 @@
+1.0.1 / 2018-02-??
+==================
+
+  * New: porter-serve --headless
+
 1.0.0 / 2018-02-11
 ==================
 
- * Breaking: renamed to porter-cli
+  * Breaking: renamed to porter-cli